# docs/swagger.yaml
definitions:
  syncer.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: 区块链浏览器后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 区块链浏览器
  version: "1.0"
paths:
  /addresses:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址列表
      tags:
      - 账户地址
  /addresses/{name}:
    get:
      consumes:
      - application/json
      description: 获取账户地址信息
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址信息
      tags:
      - 账户地址
  /addresses/{name}/assets:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址创建的相关资产
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址相关资产
      tags:
      - 账户地址
  /addresses/{name}/contractcodes:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址提交的相关合约代码
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 获取账户地址相关交易
      tags:
      - 账户地址
  /addresses/{name}/contracts:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址创建的相关合约
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址相关合约
      tags:
      - 账户地址
  /addresses/{name}/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取账户地址相关交易
      parameters:
      - description: 账户地址
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 账户地址相关交易
      tags:
      - 账户地址
  /assets:
    get:
      consumes:
      - application/json
      description: 分页获取资产列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":!200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 资产列表
      tags:
      - 资产
  /assets/{name}:
    get:
      consumes:
      - application/json
      description: 获取资产信息
      parameters:
      - description: 资产名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 资产信息
      tags:
      - 资产
  /assets/{name}/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取资产相关交易
      parameters:
      - description: 资产名称
        in: path
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 个数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":"","detail"：""}
          schema:
            $ref: '#/definitions/syncer.Response'
      summary: 资产相关相关交易
      tags:
      - 资产
...
swagger: "2.0"